@@ -1,8 +1,6 @@
 module github.com/dakusui/jqplusplus
 
-go 1.23.0
-
-toolchain go1.24.5
+go 1.24.5
 
 require (
 	github.com/BurntSushi/toml v1.6.0
